+++ smawg/tests/test_main.py
@@ -14,6 +14,7 @@ class TestMain(unittest.TestCase):
             ["python3", "-m", "smawg", "--help"],
             ["python3", "-m", "smawg", "schema"],
             ["python3", "-m", "smawg", "play", "--help"],
+            ["python3", "-m", "smawg", "viz", "--help"],
             ["python3", "-m", "smawg.cli", "--help"],
             ["python3", "-m", "smawg.viz", "--help"],
         ]